+++ lib/supabase/schema.ts
